Introduce blocking label validation
When a PR is created, labeled, unlabeled this new webhook will check for a list of blocking labels.
If one blocking label is found in the given PR, it'll update the PR status as failed.
Otherwise it'll validate the PR.

If no labels are defined in the PR, it'll block the PR too.
If no blocking labels are defined in the webhook, it'll always allow the PR.